+++ corehq/ex-submodules/pillowtop/listener.py
@@ -26,7 +26,7 @@ from pillowtop.dao.couch import CouchDocumentStore
 from pillowtop.es_utils import completely_initialize_pillow_index, doc_exists
 from pillowtop.feed.couch import CouchChangeFeed
 from pillowtop.logger import pillow_logging
-from pillowtop.pillow.interface import PillowBase
+from pillowtop.pillow.interface import PillowBase, handle_pillow_error
 from pillowtop.utils import prepare_bulk_payloads
 
 try:
@@ -622,22 +622,3 @@ class SQLPillow(SQLPillowMixIn, BasicPillow):
     def __init__(self):
         checkpoint = get_default_django_checkpoint_for_legacy_pillow_class(self.__class__)
         super(SQLPillow, self).__init__(checkpoint=checkpoint)
-
-
-def handle_pillow_error(pillow, change, exception):
-    try:
-        meta = pillow.get_couch_db().show('domain_shows/domain_date', change['id'])
-    except ResourceNotFound:
-        # Show function does not exist
-        meta = None
-    error = PillowError.get_or_create(change, pillow, change_meta=meta)
-    error.add_attempt(exception, sys.exc_info()[2])
-    error.save()
-    pillow_logging.exception(
-        "[%s] Error on change: %s, %s. Logged as: %s" % (
-            pillow.get_name(),
-            change['id'],
-            exception,
-            error.id
-        )
-    )
